+++ dantro/data_loaders/load_hdf5.py
@@ -1,122 +1,20 @@
-"""This module implements loaders mixin classes for use with the DataManager.
+"""Implements loading of Hdf5 files into the dantro data tree"""
 
-All these mixin classes should follow the pattern:
-class LoadernameLoaderMixin:
-
-    @add_loader(TargetCls=TheTargetContainerClass)
-    def _load_loadername(filepath: str, *, TargetCls):
-        # ...
-        return TargetCls(...)
-
-Each `_load_loadername` method gets supplied with path to a file and the
-`TargetCls` argument, which can be called to create an object of the correct
-type and name.
-By default, and to decouple the loader from the container, it should be
-considered to be a staticmethod; in other words: the first positional argument
-should _not_ be `self`!
-If `self` is required, `omit_self=False` may be given to the decorator.
-"""
-
-import warnings
 import logging
-
-import pickle as pkl
 
 import numpy as np
 import h5py as h5
 
-from dantro.base import BaseDataGroup, BaseDataContainer
-from dantro.container import ObjectContainer, MutableMappingContainer, NumpyDataContainer
-from dantro.group import OrderedDataGroup
-from dantro.proxy import Hdf5DataProxy
-import dantro.tools as tools
+from ..base import BaseDataGroup, BaseDataContainer
+from ..containers import NumpyDataContainer
+from ..groups import OrderedDataGroup
+from ..proxy import Hdf5DataProxy
+from ..tools import fill_line
+from ._tools import add_loader
 
 # Local constants
 log = logging.getLogger(__name__)
 
-# -----------------------------------------------------------------------------
-# Decorator to ensure correct loader function signature
-
-def add_loader(*, TargetCls, omit_self: bool=True):
-    """This decorator should be used to specify loader functions.
-    
-    Args:
-        TargetCls: Description
-        omit_self (bool, optional): If True (default), the decorated method
-            will not be supplied with the `self` object instance
-    """
-    def load_func_decorator(func):
-        """This decorator sets the load function's `TargetCls` attribute."""
-        def load_func(instance, *args, **kwargs):
-            """Calls the load function, either as with or without `self`."""
-            if omit_self:
-                return func(*args, **kwargs)
-            # not as static method
-            return func(instance, *args, **kwargs)
-
-        # Set the target class as function attribute 
-        load_func.TargetCls = TargetCls
-        return load_func
-    return load_func_decorator
-
-# -----------------------------------------------------------------------------
-
-class YamlLoaderMixin:
-    """Supplies functionality to load yaml files in the data manager"""
-
-    @add_loader(TargetCls=MutableMappingContainer)
-    def _load_yaml(filepath: str, *, TargetCls: type) -> MutableMappingContainer:
-        """Load a yaml file from the given path and creates a container to
-        store that data in.
-        
-        Args:
-            filepath (str): Where to load the yaml file from
-            TargetCls (type): The class constructor
-        
-        Returns:
-            MutableMappingContainer: The loaded yaml file as a container
-        """
-        # Load the dict
-        d = tools.load_yml(filepath)
-
-        # Populate the target container with the data
-        return TargetCls(data=d, attrs=dict(filepath=filepath))
-
-    # Also make available under `yml`
-    _load_yml = _load_yaml
-
-# -----------------------------------------------------------------------------
-
-class PickleLoaderMixin:
-    """Supplies functionality to load pickled python objects"""
-
-    # Define the load function such that it can be set in a flexible way
-    _PICKLE_LOAD_FUNC = pkl.load
-
-    @add_loader(TargetCls=ObjectContainer, omit_self=False)
-    def _load_pickle(self, filepath: str, *, TargetCls: type, **pkl_kwargs) -> ObjectContainer:
-        """Load a pickled object.
-
-        This uses the load function defined under the _PICKLE_LOAD_FUNC class
-        variable, which defaults to the pickle.load function.
-        
-        Args:
-            filepath (str): Where the pickle-dumped file is located
-            TargetCls (type): The class constructor
-            **pkl_kwargs: Passed on to the load function
-        
-        Returns:
-            ObjectContainer: The unpickled file, stored in a dantro container
-        """
-        # Open file in binary mode and unpickle with the given load function
-        with open(filepath, mode='rb') as f:
-            obj = self._PICKLE_LOAD_FUNC(f, **pkl_kwargs)
-
-        # Populate the target container with the object
-        return TargetCls(data=obj, attrs=dict(filepath=filepath))
-
-    # Also make available under `pkl`
-    _load_pkl = _load_pickle
 
 # -----------------------------------------------------------------------------
 
@@ -192,7 +90,7 @@ class Hdf5LoaderMixin:
         """
         
 
-        def recursively_load_hdf5(src, target: BaseDataGroup, *, load_as_proxy: bool, lower_case_keys: bool, enable_mapping: bool, GroupCls: BaseDataGroup, DsetCls: BaseDataContainer, GroupMap: dict, DsetMap: dict, map_attr: str):
+        def recursively_load_hdf5(src, target: BaseDataGroup, *, load_as_proxy: bool, lower_case_keys: bool, enable_mapping: bool, DsetCls: BaseDataContainer, GroupMap: dict, DsetMap: dict, map_attr: str):
             """Recursively loads the data from the source hdf5 file into the 
             target DataGroup object.
             If given, each group or dataset is checked whether an attribute
@@ -222,26 +120,26 @@ class Hdf5LoaderMixin:
                             log.warning("Could not find a mapping from map "
                                         "attribute %s='%s' to a DataGroup "
                                         "class. Available keys: %s. Falling "
-                                        "back to default class %s...",
+                                        "back to default class ...",
                                         map_attr, attrs[map_attr],
                                         ", ".join([k
-                                                   for k in GroupMap.keys()]),
-                                        GroupCls.__name__)
-                            _GroupCls = GroupCls
+                                                   for k in GroupMap.keys()]))
+                            _GroupCls = None
                     
                     else:
-                        # Use the default
-                        _GroupCls = GroupCls
+                        # Use the default of the target container
+                        _GroupCls = None
 
                     # Create and add the group, passing the attributes
-                    target.add(_GroupCls(name=key, attrs=attrs))
+                    target.new_group(path=key, Cls=_GroupCls,
+                                     attrs=attrs)
 
                     # Continue recursion
                     recursively_load_hdf5(obj, target[key],
                                           load_as_proxy=load_as_proxy,
                                           lower_case_keys=lower_case_keys,
                                           enable_mapping=enable_mapping,
-                                          GroupCls=GroupCls, DsetCls=DsetCls,
+                                          DsetCls=DsetCls,
                                           GroupMap=GroupMap, DsetMap=DsetMap,
                                           map_attr=map_attr)
 
@@ -253,7 +151,7 @@ class Hdf5LoaderMixin:
                     # Progress information on loading this dataset
                     if plvl >= 2:
                         line = fstr2.format(name=target.name, key=key, obj=obj)
-                        print(tools.fill_line(line), end="\r")
+                        print(fill_line(line), end="\r")
 
                     if load_as_proxy:
                         # Instantiate a proxy object
@@ -287,7 +185,8 @@ class Hdf5LoaderMixin:
                         _DsetCls = DsetCls
 
                     # Now create and add the dataset, passing data and attrs
-                    target.add(_DsetCls(name=key, data=data, attrs=attrs))
+                    target.new_container(path=key, Cls=_DsetCls,
+                                         data=data, attrs=attrs)
 
                 else:
                     raise NotImplementedError("Object {} is neither a dataset "
@@ -308,7 +207,6 @@ class Hdf5LoaderMixin:
 
         # Get the classes to use for groups and/or containers
         DsetCls = self._HDF5_DSET_DEFAULT_CLS
-        GroupCls = type(root)
 
         # Determine from which attribute to read the mapping
         if not map_from_attr:
@@ -331,7 +229,7 @@ class Hdf5LoaderMixin:
             if plvl >= 1:
                 # Print information on the level of this file
                 line = fstr1.format(name=root.name, file=filepath)
-                print(tools.fill_line(line), end="\r")
+                print(fill_line(line), end="\r")
 
             # Load the file level attributes, manually re-creating the dict
             root.attrs = {k:v for k, v in h5file.attrs.items()}
@@ -341,7 +239,7 @@ class Hdf5LoaderMixin:
                                   load_as_proxy=load_as_proxy,
                                   lower_case_keys=lower_case_keys,
                                   enable_mapping=enable_mapping,
-                                  GroupCls=GroupCls, DsetCls=DsetCls,
+                                  DsetCls=DsetCls,
                                   GroupMap=self._HDF5_GROUP_MAP,
                                   DsetMap=self._HDF5_DSET_MAP,
                                   map_attr=map_from_attr)