+++ dantro/data_mngr.py
@@ -9,9 +9,9 @@ import logging
 import warnings
 from typing import Union, Callable, List, Tuple
 
-from dantro.base import PATH_JOIN_CHAR, BaseDataContainer, BaseDataGroup
-from dantro.group import OrderedDataGroup
-import dantro.tools as tools
+from .base import PATH_JOIN_CHAR, BaseDataContainer, BaseDataGroup
+from .groups import OrderedDataGroup
+from .tools import fill_line, clear_line, recursive_update, load_yml
 
 # Local constants
 log = logging.getLogger(__name__)
@@ -129,11 +129,11 @@ class DataManager(OrderedDataGroup):
             # Assume this is the path to a configuration file and load it
             log.debug("Loading the default load config from a path:\n  %s",
                       load_cfg)
-            load_cfg = tools.load_yml(load_cfg)
+            load_cfg = load_yml(load_cfg)
 
         # If given, use it to recursively update the base
         if load_cfg:
-            self.load_cfg = tools.recursive_update(self.load_cfg, load_cfg)
+            self.load_cfg = recursive_update(self.load_cfg, load_cfg)
 
         # Create default groups, as specified in the _DEFAULT_GROUPS class
         # variable and the create_groups argument
@@ -274,7 +274,7 @@ class DataManager(OrderedDataGroup):
 
         if update_load_cfg:
             # Recursively update with the given keywords
-            load_cfg = tools.recursive_update(load_cfg, update_load_cfg)
+            load_cfg = recursive_update(load_cfg, update_load_cfg)
             log.debug("Updated the load configuration.")
 
         log.info("Loading %d data entries ...", len(load_cfg))
@@ -342,6 +342,10 @@ class DataManager(OrderedDataGroup):
                     With *_nowarn values, no warning is given if an entry
                     already existed.
                     Note that this is ignored when `load_as_attr` is given.
+                unpack_data (bool, optional): If True, and load_as_attr is
+                    active, not the DataContainer or DataGroup itself will be
+                    stored in the attribute, but the content of its .data
+                    attribute.
                 progress_indicator (bool): Whether to print a progress
                     indicator or not. Default: True
                 parallel (bool): If True, data is loaded in parallel.
@@ -446,7 +450,7 @@ class DataManager(OrderedDataGroup):
     # . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
     # Helpers for loading and storing data
 
-    def _load(self, *, target_path: str, loader: str, glob_str: Union[str, List[str]], load_as_attr: Union[str, None], ignore: List[str]=None, required: bool=False, path_regex: str=None, exists_action: str='raise', progress_indicator: bool=True, parallel: bool=False, **loader_kwargs) -> None:
+    def _load(self, *, target_path: str, loader: str, glob_str: Union[str, List[str]], load_as_attr: Union[str, None], ignore: List[str]=None, required: bool=False, path_regex: str=None, exists_action: str='raise', unpack_data: bool=False, progress_indicator: bool=True, parallel: bool=False, **loader_kwargs) -> None:
         """Helper function that loads a data entry to the specified path.
         
         Args:
@@ -472,6 +476,9 @@ class DataManager(OrderedDataGroup):
                 With *_nowarn values, no warning is given if an entry already
                 existed.
                 Note that this is ignored if `load_as_attr` is given.
+            unpack_data (bool, optional): If True, and load_as_attr is active,
+                not the DataContainer or DataGroup itself will be stored in
+                the attribute, but the content of its .data attribute.
             progress_indicator (bool, optional): Whether to print a progress
                 indicator or not
             parallel (bool, optional): If True, data is loaded in parallel -
@@ -687,7 +694,7 @@ class DataManager(OrderedDataGroup):
                                  "skip_nowarn, overwrite, overwrite_nowarn."
                                  "".format(exists_action))
 
-        def store(obj: Union[BaseDataGroup, BaseDataContainer], *, target_path: List[str], as_attr: Union[str, None]) -> None:
+        def store(obj: Union[BaseDataGroup, BaseDataContainer], *, target_path: List[str], as_attr: Union[str, None], unpack_data: bool) -> None:
             """Store the given `obj` at the supplied `path`.
             
             Note that this will automatically overwrite, assuming that all
@@ -730,8 +737,13 @@ class DataManager(OrderedDataGroup):
                                             "already exists in {}!"
                                             "".format(as_attr, target.logstr))
 
-                # All checks passed. Can store it now
-                target.attrs[as_attr] = obj
+                # All checks passed. Can store it now, either directly or with
+                # unpacking of its data ...
+                if not unpack_data:
+                    target.attrs[as_attr] = obj
+                else:
+                    target.attrs[as_attr] = obj.data
+                    
                 log.debug("Stored %s as attribute '%s' of %s.",
                           obj.classname, as_attr, target.logstr)
 
@@ -812,7 +824,7 @@ class DataManager(OrderedDataGroup):
         for n, file in enumerate(files):
             if progress_indicator:
                 line = "  Loading  {}/{}  ...".format(n+1, len(files))
-                print(tools.fill_line(line), end="\r")
+                print(fill_line(line), end="\r")
 
             # Prepare the target path (a list of strings)
             _target_path = prepare_target_path(target_path, filepath=file,
@@ -842,13 +854,14 @@ class DataManager(OrderedDataGroup):
                       file, _data.logstr)
             
             # If this succeeded, store the data
-            store(_data, target_path=_target_path, as_attr=load_as_attr)
+            store(_data, target_path=_target_path,
+                  as_attr=load_as_attr, unpack_data=unpack_data)
             
             # Done with this file. Go to next iteration
 
         # Clear the line to get rid of the load indicator, if there was one
         if progress_indicator:
-            tools.clear_line()
+            clear_line()
 
         # Done
         log.debug("Finished loading %d files.", len(files))
@@ -936,8 +949,7 @@ class DataManager(OrderedDataGroup):
         if path[0] not in base_group:
             log.debug("Creating group '%s' in %s ...",
                       path[0], base_group.logstr)
-            grp = GroupCls(name=path[0])
-            base_group.add(grp)
+            base_group.new_group(path[0])
 
         # path[0] is now created
         # Check whether to continue recursion